split the code in flowby into scripts "flowby", "flowbyactivity", "flowbysector", "flowbyseries"
