+++ flowsa/data_source_scripts/EPA_SIT.py
@@ -9,7 +9,7 @@ data directory. Parses EPA SIT data to flowbyactivity format.
 import pandas as pd
 import os
 from flowsa.settings import externaldatapath
-from flowsa.flowby import FlowByActivity
+from flowsa import FlowByActivity
 from flowsa.flowbyfunctions import assign_fips_location_system, \
     load_fba_w_standardized_units
 from flowsa.flowsa_log import log
