+++ flowsa/flowbyseries.py
@@ -1,0 +1,46 @@
+import pandas as pd
+
+from flowsa import FlowByActivity, FlowBySector
+from flowsa.flowby import _FlowBy
+
+
+# The three classes extending pd.Series, together with the _constructor...
+# methods of each class, are required for allowing pandas methods called on
+# objects of these classes to return objects of these classes, as desired.
+# For more information, see
+# https://pandas.pydata.org/docs/development/extending.html
+
+class _FlowBySeries(pd.Series):
+    _metadata = [*_FlowBy()._metadata]
+
+    @property
+    def _constructor(self) -> '_FlowBySeries':
+        return _FlowBySeries
+
+    @property
+    def _constructor_expanddim(self) -> '_FlowBy':
+        return _FlowBy
+
+
+class _FBASeries(pd.Series):
+    _metadata = [*FlowByActivity()._metadata]
+
+    @property
+    def _constructor(self) -> '_FBASeries':
+        return _FBASeries
+
+    @property
+    def _constructor_expanddim(self) -> 'FlowByActivity':
+        return FlowByActivity
+
+
+class _FBSSeries(pd.Series):
+    _metadata = [*FlowBySector()._metadata]
+
+    @property
+    def _constructor(self) -> '_FBSSeries':
+        return _FBSSeries
+
+    @property
+    def _constructor_expanddim(self) -> 'FlowBySector':
+        return FlowBySector
